Better Login Protocol

+++ src/main/java/org/n52/v3d/terraintools/pointset/ElevationServlet.java
@@ -29,24 +29,19 @@ package org.n52.v3d.terraintools.pointset;
 
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.PrintWriter;
-import java.util.ArrayList;
-import java.util.Enumeration;
 import java.util.List;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
 import org.n52.v3d.terraintools.drive.DriveSample;
 import org.n52.v3d.triturus.examples.gridding.Gridding;
-import org.n52.v3d.triturus.gisimplm.GmPoint;
 import org.n52.v3d.triturus.gisimplm.GmSimpleElevationGrid;
 import org.n52.v3d.triturus.gisimplm.IoElevationGridWriter;
 import org.n52.v3d.triturus.vgis.VgPoint;
@@ -57,45 +52,23 @@ import org.n52.v3d.triturus.vgis.VgPoint;
  */
 public class ElevationServlet extends HttpServlet {
 
-    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        RequestDispatcher rd = request.getRequestDispatcher("elevation.html");  
-        rd.forward(request, response);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        PrintWriter out = response.getWriter();
-
-        String pointsetId = request.getParameter("pointset");
-        String elevationName = request.getParameter("elevation");
-        
-        InputStream inputStream = DriveSample.downloadFile(DriveSample.drive, pointsetId);
-
-        File pointsetFile = File.createTempFile("tmp-pointset", ".xyz");
-        OutputStream outputStream = new FileOutputStream(pointsetFile);
-        IOUtils.copy(inputStream, outputStream);
-        outputStream.close();
-
-        String pointsetPath = pointsetFile.getPath();
-        String elevationPath = pointsetFile.getParent() + pointsetFile.separator + elevationName;
-
+    protected void doGridding(String pointsetPath, String elevationPath,
+            double cellSize, short samplingMethod) {
         Gridding gridding = new Gridding();
         gridding.setInputFile(pointsetPath);
         gridding.setOutputFile(elevationPath);
+        gridding.setCellSize(cellSize);
+        gridding.setSamplingMethod(samplingMethod);
         gridding.setOutputFormat(IoElevationGridWriter.ARCINFO_ASCII_GRID);
         List<VgPoint> points = gridding.readPointCloud();
         GmSimpleElevationGrid elevGrid = gridding.performGridding(points);
         gridding.writeOutputFile(elevGrid);
+    }
 
-        File elevationFile = new File(elevationPath);
-
-        String tokenData = (String) request.getSession().getAttribute("token");
-        DriveSample driveSample = new DriveSample(DriveSample.PROJECT_FOLDER_NAME, elevationName, elevationFile, tokenData);
+    protected void writeResponse(HttpServletResponse response, DriveSample driveSample, String pointsetId)
+            throws ServletException, IOException {
         response.setContentType("text/xml");
+        PrintWriter out = response.getWriter();
         out.println("<?xml version='1.0' encoding=\"UTF-8\" standalone=\"no\" ?>");
         //out.println("<?xml-stylesheet type=\"text/css\" href=\"terrainTools-style.css\"?>");
         out.println("<terrainToolsResponse>");
@@ -105,10 +78,55 @@ public class ElevationServlet extends HttpServlet {
         out.println("  <pointsetId>" + pointsetId + "</pointsetId>");
         out.println("  <elevationId>" + driveSample.getObjectId() + "</elevationId>");
         out.println("</terrainToolsResponse>");
+    }
 
-        request.getSession().setAttribute("elevationId", driveSample.getObjectId());
-        pointsetFile.delete();
-        elevationFile.delete();
+    protected File makeTemporaryFile(InputStream inputStream) throws IOException {
+        File pointsetFile = File.createTempFile("tmp-pointset", ".xyz");
+        OutputStream outputStream = new FileOutputStream(pointsetFile);
+        IOUtils.copy(inputStream, outputStream);
+        outputStream.close();
+        return pointsetFile;
+    }
+
+    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
+            throws ServletException, IOException {
+        response.setContentType("text/html;charset=UTF-8");
+        RequestDispatcher rd = request.getRequestDispatcher("elevation.html");
+        rd.forward(request, response);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response)
+            throws ServletException, IOException {
+        response.setContentType("text/html;charset=UTF-8");
+
+        String requestType = request.getParameter("request");
+        String pointsetId = request.getParameter("pointset");
+        String elevationName = request.getParameter("elevation");
+        short samplingMethod = Short.parseShort(request.getParameter("method"));
+        double cellSize = Double.parseDouble(request.getParameter("cellsize"));
+
+        if (requestType.equalsIgnoreCase("newElevationGrid")) {
+
+            InputStream inputStream = DriveSample.downloadFile(DriveSample.drive, pointsetId);
+
+            File pointsetFile = makeTemporaryFile(inputStream);
+
+            String pointsetPath = pointsetFile.getPath();
+            String elevationPath = pointsetFile.getParent() + pointsetFile.separator + elevationName;
+
+            doGridding(pointsetPath, elevationPath, cellSize, samplingMethod);
+
+            File elevationFile = new File(elevationPath);
+
+            String tokenData = (String) request.getSession().getAttribute("token");
+            DriveSample driveSample = new DriveSample(DriveSample.PROJECT_FOLDER_NAME, elevationName, elevationFile, tokenData);
+            writeResponse(response, driveSample, pointsetId);
+
+            pointsetFile.delete();
+            elevationFile.delete();
+        }
+
     }
 
     @Override
